TIP-860: Move Media related api tests and Controller
