+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/controllers.yml
@@ -1,5 +1,6 @@
 parameters:
     pim_enrich.controller.category_tree.class: Akeneo\Pim\Enrichment\Bundle\Controller\Ui\CategoryTreeController
+    pim_api.controller.media_file.class: Akeneo\Pim\Enrichment\Bundle\Controller\ExternalApi\MediaFileController
 
 services:
     pim_pdf_generator.controller.product:
@@ -280,3 +281,25 @@ services:
             - '@oro_security.security_facade'
             - 'pim_enrich_product_categories_view'
             - '@@AkeneoPimEnrichment/Product/listCategories.json.twig'
+
+    pim_api.controller.media_file:
+        class: '%pim_api.controller.media_file.class%'
+        arguments:
+        - '@pim_api.repository.media_file'
+        - '@pim_external_api_serializer'
+        - '@pim_api.pagination.parameter_validator'
+        - '@pim_api.pagination.offset_hal_paginator'
+        - '@akeneo_file_storage.file_storage.filesystem_provider'
+        - '@akeneo_file_storage.file_storage.file.streamed_file_fetcher'
+        - '@pim_api.repository.product'
+        - '@pim_catalog.updater.product'
+        - '@pim_catalog.saver.product'
+        - '@pim_catalog.validator.product'
+        - '@akeneo_file_storage.saver.file'
+        - '@akeneo_file_storage.file_storage.file.file_storer'
+        - '@akeneo_file_storage.remover.file'
+        - '@router'
+        - '@pim_catalog.repository.product_model'
+        - '@pim_api.updater.product_model'
+        - '@pim_catalog.saver.product_model'
+        - '%pim_api.configuration%'
